// pages/transfer-management/types.ts
export interface Employee {
  id: string;
  name: string;
  department: string;
  probationEndDate?: string;
  actualRegularizationDate?: string;
  position: string;
  tags: string[];
  employeeType?: string;
  onboardingDate: string;
  phone: string;
  email?: string;
  idNumber?: string;
  registrationStatus?: 'pending_submit' | 'abandoned' | 'passed' | 'not_sent';
  registrationName?: string;
  approvalStatus?: 'not_started' | 'passed' | 'approving' | 'rejected' | 'abandoned' | 'pending_confirm';
  confirmStatus?: 'pending_confirm' | 'confirmed';
  source?: string;
  contractCompany?: string;
  abandonTime?: string;
  abandonReason?: string;
  archiveTime?: string;
  jobId?: string;
  description?: string;
  newDepartment?: string;
  originalDepartment?: string;
  newPosition?: string;
  originalPosition?: string;
  transferDate?: string;
  transferType?: string;
  transferStatus?: string;
  transferReason?: string;
  resignationDate?: string;
  lastWorkingDate?: string;
  resignationReason?: string;
  resignationType?: string;
  resignationStatus?: 'pending' | 'resigned' | 'abandoned';
  primaryResignationReason?: string;
  resignationDescription?: string;
  planStopDate?: string;
  submissionTime?: string;
}